Adding support for git version 1.7.3.4 status line
I couldn't find which version of the git client changed the
status message, but the client on my linux box gave the
text "Changed but not updated" instead of the text
"Changes not staged" for unstaged changes. Adding this to the
doubletime theme to support the pretty red status indicator.